+++ src/main/java/com/sazaxa/shipmentapi/excel/shipping/ShippingService.java
@@ -20,7 +20,7 @@ import static com.sazaxa.shipmentapi.excel.ExcelExtension.XLSX;
 @Service
 public class ShippingService {
 
-    private final int MAX_WEIGHT = 30000;
+    private final double MAX_WEIGHT = 30000;
     private final ShippingRepository shippingRepository;
     private final ShippingRepositorySupport shippingRepositorySupport;
 
@@ -57,7 +57,7 @@ public class ShippingService {
     //국가와 무게에 따른 배송비 측정
     public double getPrice(ShippingRequestDto requestDto){
         String country = requestDto.getCountry();
-        int weight = (int) (500 * Math.ceil(requestDto.getWeight() / 500));
+        double weight = 500 * Math.ceil(requestDto.getWeight() / 500);
 
         //최대값 확인
         if (weight > MAX_WEIGHT){ weight = MAX_WEIGHT; }
